Added Log.h to the project. Fixed warning in Debugger_Display.cpp. Changed printing timeout to 10 seconds. (No rush to type PR#0!) Added SVN ignores for Release, Debug, *.user, *.ncb, *.suo.

+++ AppleWin/source/Debugger_Display.cpp
@@ -1444,7 +1444,7 @@ int GetDisassemblyLine( WORD nBaseAddress, DisasmLine_t & line_ )
 	}
 
 	int nSpaces = strlen( line_.sOpCodes );
-    while (nSpaces < nMinBytesLen)
+    while (nSpaces < (int)nMinBytesLen)
 	{
 		strcat( line_.sOpCodes, " " );
 		nSpaces++;
